+++ postgresql_audit/__init__.py
@@ -14,4 +14,4 @@ from .migrations import (  # noqa
     rename_table
 )
 
-__version__ = '0.7.0'
+__version__ = '0.8.0'